refactor(Context): change name saveContext to persistContext
